requirements.txt
requests
google-generativeai
ttkthemes
python-dotenv